requirements.txt
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.3.2
pandas==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
